+++ app/src/main/AndroidManifest.xml
@@ -6,37 +6,32 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
+        android:name=".InitRealm"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme"
-        android:name=".InitRealm"
         tools:ignore="AllowBackup,GoogleAppIndexingWarning">
-
-        <activity android:name=".PostActivity"/>
-
+        <activity android:name=".CheckGrowthActivity"></activity>
+        <activity android:name=".PostActivity" />
         <activity
             android:name=".MainActivity"
             android:label="@string/title_activity_main"
-            android:theme="@style/AppTheme.NoActionBar">
-        </activity>
+            android:theme="@style/AppTheme.NoActionBar"></activity>
         <activity android:name=".ResetPasswordMailActivity" />
-
-        <activity android:name=".SignUpActivity" />
-
-        <!-- ログイン画面 -->
+        <activity android:name=".SignUpActivity" /> <!-- ログイン画面 -->
         <activity
             android:name=".SignInActivity"
-            android:theme="@style/AppTheme.NoActionBar" >
+            android:theme="@style/AppTheme.NoActionBar">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
+
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
-        <activity android:name=".StartUpActivity"/>
+        <activity android:name=".StartUpActivity" />
 
         <meta-data
             android:name="preloaded_fonts"
